+++ src/engine/lineparams.jl
@@ -284,3 +284,37 @@ inductance(lp::LineParameters, args...) = L(lp, args...)
 conductance(lp::LineParameters, args...) = G(lp, args...)
 susceptance(lp::LineParameters, args...) = B(lp, args...)
 capacitance(lp::LineParameters, args...) = C(lp, args...)
+
+_line_component_values(::Val{:R}, object, frequencies) = R(object)
+_line_component_values(::Val{:X}, object, frequencies) = X(object)
+_line_component_values(::Val{:G}, object, frequencies) = G(object)
+_line_component_values(::Val{:B}, object, frequencies) = B(object)
+_line_component_values(::Val{:Z_re}, object, frequencies) = R(object)
+_line_component_values(::Val{:Z_im}, object, frequencies) = X(object)
+_line_component_values(::Val{:Z_abs}, object, frequencies) = abs.(Z(object))
+function _line_component_values(::Val{:Z_angle}, object, frequencies)
+    angle.(Z(object)) .* (180 / π)
+end
+_line_component_values(::Val{:Y_re}, object, frequencies) = G(object)
+_line_component_values(::Val{:Y_im}, object, frequencies) = B(object)
+_line_component_values(::Val{:Y_abs}, object, frequencies) = abs.(Y(object))
+function _line_component_values(::Val{:Y_angle}, object, frequencies)
+    angle.(Y(object)) .* (180 / π)
+end
+
+_line_component_values(::Val{:L}, parameters::LineParameters, frequencies) = L(parameters)
+_line_component_values(::Val{:C}, parameters::LineParameters, frequencies) = C(parameters)
+
+function _frequency_component_values(component::Symbol, reactive, frequencies)
+    any(iszero, frequencies) && throw(
+        DomainError(frequencies, "$component is undefined at zero frequency"),
+    )
+    return reactive ./ reshape(2π .* frequencies, 1, 1, :)
+end
+
+function _line_component_values(::Val{:L}, object::SeriesImpedance, frequencies)
+    _frequency_component_values(:L, X(object), frequencies)
+end
+function _line_component_values(::Val{:C}, object::ShuntAdmittance, frequencies)
+    _frequency_component_values(:C, B(object), frequencies)
+end